Add variable energy offset
A constant energy offset existed. Because of changes in mirror M1
configuration in beamline, the beam is not impacting on the same
place of the gratting in the monochromator. This is causing the
energy to be modified by a variable energy offset, following the
equation of a straight line.

Add support for this variable offset to the controllers, by
removing offsetEnergy and by adding two coefficients a and b
in order to deal with the straight line equation.