Remove values from last headers of RON
Do not use the value of the last header of RON gr_pitch, and
do not use the values of the two last headers of the RON m3_pitch,
because they do not read correctly the homing signal.
The signal is not reaching the 0uA.

The problem should be corrected at the HW level, before integrating
again the values of these headers of the RON, in the mean values
calculations.